move class to file
